Split compare page up to keep sidebar while loading
